+++ changelog.md
@@ -5,7 +5,10 @@
 > \[!IMPORTANT\]
 > This version is not released yet and is under active development.
 
-- Build binaries without Poetry.
+- Remove support for Poetry-based projects. All Python projects are expected to follow [standard `pyproject.toml` conventions](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/).
+- Run all Python-based commands via `uv run` and `uv tool run`.
+- Remove dependency on `poetry` package.
+- Add new dependencies on `build`, `packaging` and `pyproject-metadata`.
 
 ## [3.5.11 (2024-06-22)](https://github.com/kdeldycke/workflows/compare/v3.5.10...v3.5.11)